backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyJWT>=2.8.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
